pyproject.toml
[tool.poetry]
name = "learnify"
version = "0.1.0"
description = ""
authors = ["pecheritsa <pecheritsa.roman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
black = "^24.10.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
